[ALPS04575770] FPSGO: fix coverity
[Detail] fix coverity
Invalid type in argument to printf
format specifier:5459533

MTK-Commit-Id: 3dce4934571695bd488f7aeb5b5b6a3c272da420

CR-Id: ALPS04575770
Feature: FPSGO

+++ drivers/misc/mediatek/performance/fpsgo_v3/fbt/src/mini_top.c
@@ -389,7 +389,7 @@ end_sched_nomi:
 	ret = keep ? 0 : -EAGAIN;
 	if (mr->ever)
 		minitop_trace(" %5d %3llu%% end debnc heavy",
-			      mr->tid, (int)mr->ratio);
+			      mr->tid, mr->ratio);
 
 err_cleanup:
 	*heavy = 0;
